Refactor upgrade tracking and display in player inventory
Replaces upgrade icon mapping in Player with UpgradeBackpackItem to track both image and count. Updates upgrade acquisition logic to increment counts, and modifies UI components to display upgrade quantities. Adjusts enemy chest drop chance to use Double for finer control. GameOverScreen now shows collected upgrades and their counts. Minor tweaks to upgrade effects and UI popup behavior.

+++ src/main/kotlin/com/glycin/intelli25/managers/EnemyManager.kt
@@ -92,7 +92,7 @@ class EnemyManager(
     fun damage(enemy: Enemy, damage: Int) {
         enemy.currentHp -= damage
         if(enemy.currentHp <= 0){
-            if(Random.nextInt(100) <= enemy.chestDropChance) {
+            if(Random.nextDouble(100.0) <= enemy.chestDropChance) {
                 chestCounter++
                 treasureMap[chestCounter] = TreasureChest(
                     id = chestCounter,
